Extract workspace terminal APIs
